Update Namespace for authorize
